Phase 2: start the passage/ package split with the /voice page
Extract the most self-contained slice of TranslationUI.py first: the
/voice page (recorder UI, its ~450-line head-injected JS, and
api_voice_translate) moves into passage/ui/voice_page.py as a
VoicePageMixin. TranslationUI still inherits it, so shared state
(self.backend, _check_api_access, _inject_theme) needs no redesign -
this is the safe way to split a stateful NiceGUI controller class
without a bigger rearchitecture.

LANGUAGES and _log_event move to passage/ui/common.py so voice_page.py
doesn't need a circular import back to TranslationUI.py.

TranslationUI.py: 1672 -> 1225 lines. The workspace/segment-editor
methods stay put for now - they share heavy self.* state and a
mixin-only split there is lower value than the remaining phases;
noted as deferred, not abandoned, in PASSAGE_PLAN.md.

Also: deleted Multilingual.png, an unreferenced 2MB stray asset.

Verified: 71/71 pytest (3 tests updated to grep the new file for the
JS they check), live browser 7/7 feature pass including the moved
/voice route, live voice HTTP round-trip against the real API.
